requirements.txt
nonebot2[fastapi]>=2.0.0
nonebot-adapter-onebot>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
